updating module name for consistency and fixing secure_open issue
